Begin move Python files to `core` folder
